Moving owned and unowned products into a hook
